@@ -9,25 +9,31 @@ using System.Threading.Tasks;
 using AzureNamingTool.Services;
 using AzureNamingTool.Attributes;
 
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
-
 namespace AzureNamingTool.Controllers
 {
+    /// <summary>
+    /// Controller for importing and exporting configuration data.
+    /// </summary>
     [Route("api/[controller]")]
     [ApiController]
     [ApiKey]
     public class ImportExportController : ControllerBase
     {
+        /// <summary>
+        /// Response for controller functions
+        /// </summary>
+        ServiceResponse serviceResponse = new();
+
         // GET: api/<ImportExportController>
         /// <summary>
         /// This function will export the current configuration data (all components) as a single JSON file. 
         /// </summary>
-        /// <returns>json - JSON configuration file</returns>
+        /// <param name="includeAdmin">Flag indicating whether to include admin data in the export</param>
+        /// <returns>JSON configuration file</returns>
         [HttpGet]
         [Route("[action]")]
         public async Task<IActionResult> ExportConfiguration(bool includeAdmin = false)
         {
-            ServiceResponse serviceResponse = new();
             try
             {
                 serviceResponse = await ImportExportService.ExportConfig(includeAdmin);
@@ -51,13 +57,12 @@ namespace AzureNamingTool.Controllers
         /// <summary>
         /// This function will import the provided configuration data (all components). This will overwrite the existing configuration data. 
         /// </summary>
-        /// <param name="configdata">ConfigurationData (json) - Tool configuration File</param>
-        /// <returns>bool - PASS/FAIL</returns>
+        /// <param name="configdata">Tool configuration file in JSON format</param>
+        /// <returns>Flag indicating whether the import was successful or not</returns>
         [HttpPost]
         [Route("[action]")]
         public async Task<IActionResult> ImportConfiguration([FromBody] ConfigurationData configdata)
         {
-            ServiceResponse serviceResponse = new();
             try
             {
                 serviceResponse = await ImportExportService.PostConfig(configdata);
