@@ -8,10 +8,12 @@ using System.Linq;
 using System.Net;
 using AzureNamingTool.Services;
 using AzureNamingTool.Attributes;
-// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace AzureNamingTool.Controllers
 {
+    /// <summary>
+    /// Controller for managing policies.
+    /// </summary>
     [Route("api/[controller]")]
     [ApiController]
     [ApiKey]
