@@ -13,8 +13,17 @@ using System.Threading.Tasks;
 
 namespace AzureNamingTool.Helpers
 {
+    /// <summary>
+    /// Helper class for file system operations.
+    /// </summary>
     public class FileSystemHelper
     {
+        /// <summary>
+        /// Reads the content of a file.
+        /// </summary>
+        /// <param name="fileName">The name of the file to read.</param>
+        /// <param name="folderName">The folder name where the file is located. Default is "settings/".</param>
+        /// <returns>The content of the file.</returns>
         public static async Task<string> ReadFile(string fileName, string folderName = "settings/")
         {
             await CheckFile(folderName + fileName);
@@ -22,6 +31,12 @@ namespace AzureNamingTool.Helpers
             return data;
         }
 
+        /// <summary>
+        /// Writes content to a file.
+        /// </summary>
+        /// <param name="fileName">The name of the file to write.</param>
+        /// <param name="content">The content to write to the file.</param>
+        /// <param name="folderName">The folder name where the file should be located. Default is "settings/".</param>
         public static async Task WriteFile(string fileName, string content, string folderName = "settings/")
         {
             await CheckFile(folderName + fileName);
@@ -45,6 +60,10 @@ namespace AzureNamingTool.Helpers
             }
         }
 
+        /// <summary>
+        /// Checks if a file exists. If not, creates an empty file.
+        /// </summary>
+        /// <param name="fileName">The name of the file to check.</param>
         public static async Task CheckFile(string fileName)
         {
             if (!File.Exists(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, fileName)))
@@ -67,11 +86,16 @@ namespace AzureNamingTool.Helpers
             }
         }
 
+        /// <summary>
+        /// Writes configuration data to a file.
+        /// </summary>
+        /// <param name="configdata">The configuration data to write.</param>
+        /// <param name="configFileName">The name of the configuration file.</param>
+        /// <returns>A message indicating the status of the operation.</returns>
         public static async Task<object> WriteConfiguation(object configdata, string configFileName)
         {
             try
             {
-
                 var options = new JsonSerializerOptions
                 {
                     PropertyNamingPolicy = JsonNamingPolicy.CamelCase
@@ -87,22 +111,23 @@ namespace AzureNamingTool.Helpers
             }
         }
 
+        /// <summary>
+        /// Resets the configuration file by copying it from the repository folder to the settings folder.
+        /// </summary>
+        /// <param name="filename">The name of the configuration file to reset.</param>
+        /// <returns>True if the configuration file was reset successfully, otherwise false.</returns>
         public static bool ResetConfiguration(string filename)
         {
             bool result = false;
             try
             {
-                // Get all the files in the repository folder
                 DirectoryInfo dirRepository = new("repository");
                 foreach (FileInfo file in dirRepository.GetFiles())
                 {
                     if (file.Name == filename)
                     {
-                        // Copy the repository file to the settings folder
                         file.CopyTo(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "settings/" + file.Name), true);
                         result = true;
-                        // Clear any cached data
-
                         break;
                     }
                 }
@@ -114,15 +139,19 @@ namespace AzureNamingTool.Helpers
             return result;
         }
 
+        /// <summary>
+        /// Migrates data from one file to another.
+        /// </summary>
+        /// <param name="sourcefileName">The name of the source file.</param>
+        /// <param name="sourcefolderName">The folder name where the source file is located.</param>
+        /// <param name="destinationfilename">The name of the destination file.</param>
+        /// <param name="destinationfolderName">The folder name where the destination file should be located.</param>
+        /// <param name="delete">Indicates whether to delete the source file after migration.</param>
         public static async Task MigrateDataToFile(string sourcefileName, string sourcefolderName, string destinationfilename, string destinationfolderName, bool delete)
         {
-            // Get the source data
             string data = await ReadFile(sourcefileName, sourcefolderName);
-
-            // Write the destination data
             await WriteFile(destinationfilename, data, destinationfolderName);
 
-            // Check if the source file should be removed (In repository and settings folders)
             if (delete)
             {
                 File.Delete(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "repository/" + sourcefileName));
