@@ -1,9 +1,10 @@
 using AzureNamingTool.Models;
-using AzureNamingTool.Pages;
+using AzureNamingTool.Components.Pages;
 using AzureNamingTool.Services;
-using AzureNamingTool.Shared;
+using AzureNamingTool.Components;
 using Blazored.Modal;
 using Blazored.Modal.Services;
+using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using System.Security.Cryptography;
@@ -11,9 +12,18 @@ using System.Text;
 
 namespace AzureNamingTool.Helpers
 {
+
+    /// <summary>
+    /// Helper class for general operations.
+    /// </summary>
     public class GeneralHelper
     {
-        //Function to get the Property Value
+        /// <summary>
+        /// Function to get the Property Value
+        /// </summary>
+        /// <param name="SourceData">The source data object</param>
+        /// <param name="propName">The name of the property</param>
+        /// <returns>The value of the property</returns>
         public static object? GetPropertyValue(object SourceData, string propName)
         {
             try
@@ -27,6 +37,12 @@ namespace AzureNamingTool.Helpers
             }
         }
 
+        /// <summary>
+        /// Encrypts a string using AES encryption.
+        /// </summary>
+        /// <param name="text">The text to encrypt</param>
+        /// <param name="keyString">The encryption key</param>
+        /// <returns>The encrypted string</returns>
         public static string EncryptString(string text, string keyString)
         {
             byte[] iv = new byte[16];
@@ -49,6 +65,12 @@ namespace AzureNamingTool.Helpers
             return Convert.ToBase64String(array);
         }
 
+        /// <summary>
+        /// Decrypts a string using AES decryption.
+        /// </summary>
+        /// <param name="cipherText">The encrypted string</param>
+        /// <param name="keyString">The decryption key</param>
+        /// <returns>The decrypted string</returns>
         public static string DecryptString(string cipherText, string keyString)
         {
             byte[] iv = new byte[16];
@@ -65,6 +87,11 @@ namespace AzureNamingTool.Helpers
             return streamReader.ReadToEnd();
         }
 
+        /// <summary>
+        /// Checks if a string is base64 encoded.
+        /// </summary>
+        /// <param name="value">The string to check</param>
+        /// <returns>True if the string is base64 encoded, otherwise false</returns>
         public static bool IsBase64Encoded(string value)
         {
             bool base64encoded = false;
@@ -80,7 +107,11 @@ namespace AzureNamingTool.Helpers
             return base64encoded;
         }
 
-
+        /// <summary>
+        /// Downloads a string from a URL using HttpClient.
+        /// </summary>
+        /// <param name="url">The URL to download from</param>
+        /// <returns>The downloaded string</returns>
         public static async Task<string> DownloadString(string url)
         {
             string data;
@@ -97,6 +128,12 @@ namespace AzureNamingTool.Helpers
             return data;
         }
 
+        /// <summary>
+        /// Normalizes a name by removing "Resource" and spaces.
+        /// </summary>
+        /// <param name="name">The name to normalize</param>
+        /// <param name="lowercase">True to convert the name to lowercase, false otherwise</param>
+        /// <returns>The normalized name</returns>
         public static string NormalizeName(string name, bool lowercase)
         {
             string newname = name.Replace("Resource", "").Replace(" ", "");
@@ -107,6 +144,11 @@ namespace AzureNamingTool.Helpers
             return newname;
         }
 
+        /// <summary>
+        /// Sets the CSS class for text based on the enabled state.
+        /// </summary>
+        /// <param name="enabled">True if the text is enabled, false if disabled</param>
+        /// <returns>The CSS class for the text</returns>
         public static string SetTextEnabledClass(bool enabled)
         {
             if (enabled)
@@ -119,9 +161,18 @@ namespace AzureNamingTool.Helpers
             }
         }
 
+        /// <summary>
+        /// Checks if an object is not null.
+        /// </summary>
+        /// <param name="obj">The object to check</param>
+        /// <returns>True if the object is not null, otherwise false</returns>
         public static bool IsNotNull([NotNullWhen(true)] object? obj) => obj != null;
 
-
+        /// <summary>
+        /// Formats a resource type string.
+        /// </summary>
+        /// <param name="type">The resource type string</param>
+        /// <returns>An array containing the formatted resource type</returns>
         public static string[] FormatResoureType(string type)
         {
             String[] returntype = new String[4];
@@ -129,7 +180,7 @@ namespace AzureNamingTool.Helpers
             // Make sure it is a full resource type name
             if (type.Contains('('))
             {
-                returntype[0] = type[..type.IndexOf("(")].Trim();
+                returntype[0] = type[..type.IndexOf('(')].Trim();
             }
             try
             {
@@ -142,14 +193,14 @@ namespace AzureNamingTool.Helpers
                         // Get all text before the dash
                         returntype[1] = returntype[0][..returntype[0].IndexOf(" -")].Trim();
                         // Get all text after the dash
-                        returntype[3] = returntype[0].Substring(returntype[0].IndexOf("-") + 1).Trim();
+                        returntype[3] = returntype[0][(returntype[0].IndexOf('-') + 1)..].Trim();
                     }
 
                     // trim any details out of the value
                     if (type.Contains('(') && type.Contains(')'))
                     {
                         {
-                            int intstart = type.IndexOf("(") + 1;
+                            int intstart = type.IndexOf('(') + 1;
                             returntype[2] = String.Concat(type[intstart..].TakeWhile(x => x != ')'));
                         }
                     }
@@ -160,6 +211,33 @@ namespace AzureNamingTool.Helpers
                 AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
             }
             return returntype;
+        }
+
+        /// <summary>
+        /// Generates a random string of specified length.
+        /// </summary>
+        /// <param name="maxLength">The maximum length of the string</param>
+        /// <param name="alphanumeric">True to include alphanumeric characters, false for lowercase alphabetic characters only</param>
+        /// <returns>The generated random string</returns>
+        public static string GenerateRandomString(int maxLength, bool alphanumeric)
+        {
+            var chars = "abcdefghijklmnopqrstuvwxyz";
+            if (alphanumeric)
+            {
+                chars = "abcdefghijklmnopqrstuvwxyz0123456789";
+            }
+
+            var Charsarr = new char[maxLength];
+            var random = new Random();
+
+            for (int i = 0; i < Charsarr.Length; i++)
+            {
+                Charsarr[i] = chars[random.Next(chars.Length)];
+            }
+
+            var result = new String(Charsarr);
+
+            return result;
         }
     }
 }
