@@ -1,8 +1,18 @@
 ﻿namespace AzureNamingTool.Models
 {
+    /// <summary>
+    /// Represents the details of an identity provider.
+    /// </summary>
     public class IdentityProviderDetails
     {
+        /// <summary>
+        /// Gets or sets the current user.
+        /// </summary>
         public string CurrentUser { get; set; } = "System";
-        public string CurrentIdentityProvider { get; set; } =  String.Empty;
+
+        /// <summary>
+        /// Gets or sets the current identity provider.
+        /// </summary>
+        public string CurrentIdentityProvider { get; set; } = String.Empty;
     }
 }
