@@ -1,8 +1,18 @@
 ﻿namespace AzureNamingTool.Models
 {
+    /// <summary>
+    /// Represents information about a theme.
+    /// </summary>
     public class ThemeInfo
     {
-        public string ThemeName { get; set; } =  String.Empty;
-        public string ThemeStyle { get; set; } =  String.Empty;
+        /// <summary>
+        /// Gets or sets the name of the theme.
+        /// </summary>
+        public string ThemeName { get; set; } = String.Empty;
+
+        /// <summary>
+        /// Gets or sets the style of the theme.
+        /// </summary>
+        public string ThemeStyle { get; set; } = String.Empty;
     }
 }
