@@ -5,10 +5,18 @@ using System;
 
 namespace AzureNamingTool.Services
 {
+    /// <summary>
+    /// Service for managing the admin settings.
+    /// </summary>
     public class AdminService
     {
         private static readonly SiteConfiguration config = ConfigurationHelper.GetConfigurationData();
 
+        /// <summary>
+        /// Updates the user password.
+        /// </summary>
+        /// <param name="password">The new password.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> UpdatePassword(string password)
         {
             ServiceResponse serviceResponse = new();
@@ -23,7 +31,7 @@ namespace AzureNamingTool.Services
                 else
                 {
                     serviceResponse.Success = false;
-                    serviceResponse.ResponseObject = "The pasword does not meet the security requirements.";
+                    serviceResponse.ResponseObject = "The password does not meet the security requirements.";
                 }
             }
             catch (Exception ex)
@@ -35,6 +43,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Generates a new API key based on the specified type.
+        /// </summary>
+        /// <param name="type">The type of API key to generate (fullaccess or readonly).</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> GenerateAPIKey(string type)
         {
             ServiceResponse serviceResponse = new();
@@ -42,7 +55,7 @@ namespace AzureNamingTool.Services
             {
                 // Set the new api key
                 Guid guid = Guid.NewGuid();
-                switch(type)
+                switch (type)
                 {
                     case "fullaccess":
                         config.APIKey = GeneralHelper.EncryptString(guid.ToString(), config.SALTKey!);
@@ -64,12 +77,18 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Updates the API key based on the specified type.
+        /// </summary>
+        /// <param name="apikey">The new API key.</param>
+        /// <param name="type">The type of API key to update (fullaccess or readonly).</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> UpdateAPIKey(string apikey, string type)
         {
             ServiceResponse serviceResponse = new();
             try
             {
-                switch(type)
+                switch (type)
                 {
                     case "fullaccess":
                         config.APIKey = GeneralHelper.EncryptString(apikey, config.SALTKey!);
@@ -93,6 +112,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Updates the identity header name.
+        /// </summary>
+        /// <param name="identityheadername">The new identity header name.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> UpdateIdentityHeaderName(string identityheadername)
         {
             ServiceResponse serviceResponse = new();
