@@ -3,8 +3,15 @@ using AzureNamingTool.Models;
 
 namespace AzureNamingTool.Services
 {
+    /// <summary>
+    /// Service for managing custom components.
+    /// </summary>
     public class CustomComponentService
     {
+        /// <summary>
+        /// Retrieves a list of custom component items.
+        /// </summary>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> GetItems()
         {
             ServiceResponse serviceResponse = new();
@@ -31,6 +38,53 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Retrieves a list of custom component items by the parent component ID.
+        /// </summary>
+        /// <param name="parentcomponetid">The ID of the parent component.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
+        public static async Task<ServiceResponse> GetItemsByParentComponentId(int parentcomponetid)
+        {
+            ServiceResponse serviceResponse = new();
+            try
+            {
+                // Get the parent component details
+                serviceResponse = await ResourceComponentService.GetItem(parentcomponetid);
+                if(serviceResponse.Success)
+                {
+                    var component = (ResourceComponent)serviceResponse.ResponseObject!;
+
+                    // Get list of items
+                    var items = await ConfigurationHelper.GetList<CustomComponent>();
+                    if (GeneralHelper.IsNotNull(items))
+                    {
+                        serviceResponse.ResponseObject = items.Where(x => x.ParentComponent == GeneralHelper.NormalizeName(component.Name, true)).OrderBy(x => x.SortOrder).ToList();
+                        serviceResponse.Success = true;
+                    }
+                    else
+                    {
+                        serviceResponse.ResponseObject = "Custom Components not found!";
+                    }
+                }
+                else
+                {
+                    serviceResponse.ResponseObject = "Resource Component not found!";
+                }
+            }
+            catch (Exception ex)
+            {
+                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                serviceResponse.Success = false;
+                serviceResponse.ResponseObject = ex;
+            }
+            return serviceResponse;
+        }
+
+        /// <summary>
+        /// Retrieves a list of custom component items by the parent type.
+        /// </summary>
+        /// <param name="parenttype">The parent type of the custom components.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> GetItemsByParentType(string parenttype)
         {
             ServiceResponse serviceResponse = new();
@@ -57,6 +111,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Retrieves a custom component item by its ID.
+        /// </summary>
+        /// <param name="id">The ID of the custom component to retrieve.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> GetItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -91,6 +150,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Posts a custom component item.
+        /// </summary>
+        /// <param name="item">The custom component item to be posted.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> PostItem(CustomComponent item)
         {
             ServiceResponse serviceResponse = new();
@@ -104,9 +168,6 @@ namespace AzureNamingTool.Services
                     return serviceResponse;
                 }
 
-                // Force lowercase on the shortname
-                item.ShortName = item.ShortName.ToLower();
-
                 // Get list of items
                 var items = await ConfigurationHelper.GetList<CustomComponent>();
                 if (GeneralHelper.IsNotNull(items))
@@ -118,7 +179,7 @@ namespace AzureNamingTool.Services
                     }
 
                     int position = 1;
-                    items = items.OrderBy(x => x.SortOrder).ToList();
+                    items = [.. items.OrderBy(x => x.SortOrder)];
 
                     if (item.SortOrder == 0)
                     {
@@ -159,10 +220,11 @@ namespace AzureNamingTool.Services
                         else
                         {
                             // Put the item at the end
+                            item.SortOrder = position;
                             items.Add(item);
                         }
                     }
-                    else
+                                    else
                     {
                         item.Id = 1;
                         item.SortOrder = 1;
@@ -191,6 +253,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Deletes a custom component by its ID.
+        /// </summary>
+        /// <param name="id">The ID of the custom component to delete.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> DeleteItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -238,6 +305,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Posts a configuration of custom components.
+        /// </summary>
+        /// <param name="items">The list of custom components to be configured.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
         public static async Task<ServiceResponse> PostConfig(List<CustomComponent> items)
         {
             ServiceResponse serviceResponse = new();
@@ -258,9 +330,6 @@ namespace AzureNamingTool.Services
                         return serviceResponse;
                     }
 
-                    // Force lowercase on the shortname
-                    item.ShortName = item.ShortName.ToLower();
-
                     item.Id = i;
                     item.SortOrder = i;
                     newitems.Add(item);
@@ -270,6 +339,74 @@ namespace AzureNamingTool.Services
                 // Write items to file
                 await ConfigurationHelper.WriteList<CustomComponent>(newitems);
                 serviceResponse.Success = true;
+            }
+            catch (Exception ex)
+            {
+                AdminLogService.PostItem(new AdminLogMessage() { Title = "ERROR", Message = ex.Message });
+                serviceResponse.ResponseObject = ex;
+                serviceResponse.Success = false;
+            }
+            return serviceResponse;
+        }
+
+        /// <summary>
+        /// Deletes custom components by parent component ID.
+        /// </summary>
+        /// <param name="componentid">The ID of the parent component.</param>
+        /// <returns>A <see cref="Task{ServiceResponse}"/> representing the asynchronous operation.</returns>
+        public static async Task<ServiceResponse> DeleteByParentComponentId(int componentid)
+        {
+            ServiceResponse serviceResponse = new();
+            try
+            {
+                // Get the resource component
+                serviceResponse = await ResourceComponentService.GetItem(componentid);
+                if (serviceResponse.Success)
+                {
+                    var component = serviceResponse.ResponseObject as ResourceComponent;
+                    if (GeneralHelper.IsNotNull(component))
+                    {
+                        // Get list of items
+                        var items = await ConfigurationHelper.GetList<CustomComponent>();
+                        if (GeneralHelper.IsNotNull(items))
+                        {
+                            // Get the custom component options
+                            List<CustomComponent> customcomponents = items.Where(x => GeneralHelper.NormalizeName(component.Name, true) == x.ParentComponent).ToList();
+                            if (GeneralHelper.IsNotNull(customcomponents))
+                            {
+                                foreach (CustomComponent customcomponent in customcomponents)
+                                {
+                                    // Remove the item from the collection
+                                    items.Remove(customcomponent);
+
+                                    // Update all the sort order values to reflect the removal
+                                    int position = 1;
+                                    foreach (CustomComponent thisitem in items.OrderBy(x => x.SortOrder).ToList())
+                                    {
+                                        thisitem.SortOrder = position;
+                                        position += 1;
+                                    }
+                                }
+
+                                // Write items to file
+                                await ConfigurationHelper.WriteList<CustomComponent>(items);
+                                serviceResponse.Success = true;
+                            }
+                            else
+                            {
+                                serviceResponse.ResponseObject = "Custom Component not found!";
+                            }
+                        }
+                        else
+                        {
+                            serviceResponse.ResponseObject = "Custom Component not found!";
+                        }
+                    }
+                    else
+                    {
+                        serviceResponse.ResponseObject = "Custom Component not found!";
+                    }
+                }
             }
             catch (Exception ex)
             {
