@@ -3,8 +3,15 @@ using AzureNamingTool.Models;
 
 namespace AzureNamingTool.Services
 {
+    /// <summary>
+    /// Service for managing resource environments.
+    /// </summary>
     public class ResourceEnvironmentService
     {
+        /// <summary>
+        /// Retrieves a list of resource environments.
+        /// </summary>
+        /// <returns>A <see cref="ServiceResponse"/> containing the list of resource environments if found, or an error message if not found.</returns>
         public static async Task<ServiceResponse> GetItems()
         {
             ServiceResponse serviceResponse = new();
@@ -31,6 +38,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Retrieves an item from the list of resource environments based on the specified ID.
+        /// </summary>
+        /// <param name="id">The ID of the item to retrieve.</param>
+        /// <returns>A <see cref="ServiceResponse"/> containing the retrieved item if found, or an error message if not found.</returns>
         public static async Task<ServiceResponse> GetItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -65,6 +77,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Posts an item to the list of resource environments.
+        /// </summary>
+        /// <param name="item">The item to add or update.</param>
+        /// <returns>A <see cref="ServiceResponse"/> indicating the success of the operation.</returns>
         public static async Task<ServiceResponse> PostItem(ResourceEnvironment item)
         {
             ServiceResponse serviceResponse = new();
@@ -77,9 +94,6 @@ namespace AzureNamingTool.Services
                     serviceResponse.ResponseObject = "Short name must be alphanumeric.";
                     return serviceResponse;
                 }
-
-                // Force lowercase on the shortname
-                item.ShortName = item.ShortName.ToLower();
 
                 // Get list of items
                 var items = await ConfigurationHelper.GetList<ResourceEnvironment>();
@@ -99,7 +113,7 @@ namespace AzureNamingTool.Services
                     }
 
                     int position = 1;
-                    items = items.OrderBy(x => x.SortOrder).ToList();
+                    items = [.. items.OrderBy(x => x.SortOrder)];
 
                     if (item.SortOrder == 0)
                     {
@@ -137,6 +151,7 @@ namespace AzureNamingTool.Services
                         else
                         {
                             // Put the item at the end
+                            item.SortOrder = position;
                             items.Add(item);
                         }
                     }
@@ -173,6 +188,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Deletes an item from the list of resource environments.
+        /// </summary>
+        /// <param name="id">The ID of the item to delete.</param>
+        /// <returns>A <see cref="ServiceResponse"/> indicating the success of the operation.</returns>
         public static async Task<ServiceResponse> DeleteItem(int id)
         {
             ServiceResponse serviceResponse = new();
@@ -220,6 +240,11 @@ namespace AzureNamingTool.Services
             return serviceResponse;
         }
 
+        /// <summary>
+        /// Posts the configuration for a list of resource environments.
+        /// </summary>
+        /// <param name="items">The list of resource environments to configure.</param>
+        /// <returns>A <see cref="ServiceResponse"/> indicating the success of the operation.</returns>
         public static async Task<ServiceResponse> PostConfig(List<ResourceEnvironment> items)
         {
             ServiceResponse serviceResponse = new();
@@ -239,9 +264,6 @@ namespace AzureNamingTool.Services
                         serviceResponse.ResponseObject = "Short name must be alphanumeric.";
                         return serviceResponse;
                     }
-
-                    // Force lowercase on the shortname
-                    item.ShortName = item.ShortName.ToLower();
 
                     item.Id = i;
                     item.SortOrder = i;
