@@ -13,6 +13,9 @@ using static System.Net.Mime.MediaTypeNames;
 
 namespace AzureNamingTool.Services
 {
+    /// <summary>
+    /// Service for managing Resource Naming Requests
+    /// </summary>
     public class ResourceNamingRequestService
     {
         /// <summary>
@@ -32,7 +35,8 @@ namespace AzureNamingTool.Services
             {
                 bool valid = true;
                 bool ignoredelimeter = false;
-                List<string[]> lstComponents = new();
+                List<string[]> lstComponents = [];
+                bool previousdelimiterappliedafter = true;
 
                 // Get the specified resource type
                 //var resourceTypes = await ConfigurationHelper.GetList<ResourceType>();
@@ -50,9 +54,8 @@ namespace AzureNamingTool.Services
                 }
 
                 // Get the components
-                ServiceResponse serviceresponse = new();
-                serviceresponse = await ResourceComponentService.GetItems(false);
-                var currentResourceComponents = serviceresponse.ResponseObject;
+                serviceResponse = await ResourceComponentService.GetItems(false);
+                var currentResourceComponents = serviceResponse.ResponseObject;
                 dynamic d = request;
 
                 string name = "";
@@ -81,7 +84,7 @@ namespace AzureNamingTool.Services
                                 }
                                 else
                                 {
-                                    value = prop.GetType().GetProperty("ShortName").GetValue(prop, null).ToLower();
+                                    value = prop.GetType().GetProperty("ShortName").GetValue(prop, null);
                                 }
 
                                 // Check if the delimeter is already ignored
@@ -94,7 +97,11 @@ namespace AzureNamingTool.Services
                                         {
                                             if (!String.IsNullOrEmpty(name))
                                             {
-                                                name += request.ResourceDelimiter.Delimiter;
+                                                // Check if the component should apply the delimiter
+                                                if ((component.ApplyDelimiterBefore) && (previousdelimiterappliedafter))
+                                                {
+                                                    name += request.ResourceDelimiter.Delimiter;
+                                                }
                                             }
                                         }
                                         else
@@ -110,7 +117,11 @@ namespace AzureNamingTool.Services
                                         // Deliemeter is valid so add it
                                         if (!String.IsNullOrEmpty(name))
                                         {
-                                            name += request.ResourceDelimiter.Delimiter;
+                                            // Check if the component should apply the delimiter
+                                            if ((component.ApplyDelimiterBefore) && (previousdelimiterappliedafter))
+                                            {
+                                                name += request.ResourceDelimiter.Delimiter;
+                                            }
                                         }
                                     }
                                 }
@@ -120,17 +131,17 @@ namespace AzureNamingTool.Services
                                 // Add property to aray for indivudal component validation
                                 if (component.Name == "ResourceType")
                                 {
-                                    lstComponents.Add(new string[] { component.Name, prop.Resource + " (" + value + ")" });
+                                    lstComponents.Add([component.Name, prop.Resource + " (" + value + ")"]);
                                 }
                                 else
                                 {
                                     if (component.Name == "ResourceInstance")
                                     {
-                                        lstComponents.Add(new string[] { component.Name, prop });
+                                        lstComponents.Add([component.Name, prop]);
                                     }
                                     else
                                     {
-                                        lstComponents.Add(new string[] { component.Name, prop.Name + " (" + value + ")" });
+                                        lstComponents.Add([component.Name, prop.Name + " (" + value + ")"]);
                                     }
                                 }
                             }
@@ -144,6 +155,7 @@ namespace AzureNamingTool.Services
                                 }
                             }
                         }
+                        previousdelimiterappliedafter = component.ApplyDelimiterAfter;
                     }
                 }
                 // Check if the required component were supplied
@@ -172,6 +184,7 @@ namespace AzureNamingTool.Services
                 // CALL VALIDATION FUNCTION
                 ValidateNameRequest validateNameRequest = new()
                 {
+                    ResourceTypeId = resourceType.Id,
                     ResourceType = resourceType.ShortName,
                     Name = name
                 };
@@ -198,12 +211,12 @@ namespace AzureNamingTool.Services
                     GeneratedName generatedName = new()
                     {
                         CreatedOn = DateTime.Now,
-                        ResourceName = name.ToLower(),
+                        ResourceName = name,
                         Components = lstComponents
                     };
                     await GeneratedNamesService.PostItem(generatedName);
                     response.Success = true;
-                    response.ResourceName = name.ToLower();
+                    response.ResourceName = name;
                     response.Message = sbMessage.ToString();
                     return response;
                 }
@@ -223,7 +236,7 @@ namespace AzureNamingTool.Services
         }
 
         /// <summary>
-        /// This function is used to generate a name by providing each componetn and the short name value. The function will validate the values to ensure they match the current configuration. 
+        /// This function is used to generate a name by providing each component and the short name value. The function will validate the values to ensure they match the current configuration. 
         /// </summary>
         /// <param name="request"></param>
         /// <returns>ResourceNameResponse - Response of name generation</returns>
@@ -238,12 +251,13 @@ namespace AzureNamingTool.Services
             {
                 bool valid = true;
                 bool ignoredelimeter = false;
-                List<string[]> lstComponents = new();
+                List<string[]> lstComponents = [];
                 ServiceResponse serviceResponse = new();
                 ResourceDelimiter resourceDelimiter = new();
                 ResourceType resourceType = new();
                 string name = "";
                 StringBuilder sbMessage = new();
+                bool previousdelimiterappliedafter = true;
 
                 // Get the current delimiter
                 serviceResponse = await ResourceDelimiterService.GetCurrentItem();
@@ -324,7 +338,7 @@ namespace AzureNamingTool.Services
                     // Make sure the passed custom component names are normalized
                     if (GeneralHelper.IsNotNull(request.CustomComponents))
                     {
-                        Dictionary<string, string> newComponents = new();
+                        Dictionary<string, string> newComponents = [];
                         foreach (var cc in request.CustomComponents)
                         {
                             string value = cc.Value;
@@ -359,132 +373,149 @@ namespace AzureNamingTool.Services
                                             {
                                                 if (!String.IsNullOrEmpty(value))
                                                 {
-                                                    // Validate that the value is a valid option for the component
-                                                    switch (component.Name.ToLower())
+                                                    // Check of the component value length is valid
+                                                    if (!ValidationHelper.CheckComponentLength(component, value))
                                                     {
-                                                        case "resourcetype":
-                                                            var types = await ConfigurationHelper.GetList<ResourceType>();
-                                                            if (GeneralHelper.IsNotNull(types))
-                                                            {
-                                                                var type = types.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(type))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceType value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourceenvironment":
-                                                            var environments = await ConfigurationHelper.GetList<ResourceEnvironment>();
-                                                            if (GeneralHelper.IsNotNull(environments))
-                                                            {
-                                                                var environment = environments.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(environment))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceEnvironment value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourcelocation":
-                                                            var locations = await ConfigurationHelper.GetList<ResourceLocation>();
-                                                            if (GeneralHelper.IsNotNull(locations))
-                                                            {
-                                                                var location = locations.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(location))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceLocation value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourceorg":
-                                                            var orgs = await ConfigurationHelper.GetList<ResourceOrg>();
-                                                            if (GeneralHelper.IsNotNull(orgs))
-                                                            {
-                                                                var org = orgs.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(org))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("Resource Type value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourceprojappsvc":
-                                                            var projappsvcs = await ConfigurationHelper.GetList<ResourceProjAppSvc>();
-                                                            if (GeneralHelper.IsNotNull(projappsvcs))
-                                                            {
-                                                                var projappsvc = projappsvcs.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(projappsvc))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceProjAppSvc value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourceunitdept":
-                                                            var unitdepts = await ConfigurationHelper.GetList<ResourceUnitDept>();
-                                                            if (GeneralHelper.IsNotNull(unitdepts))
-                                                            {
-                                                                var unitdept = unitdepts.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(unitdept))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceUnitDept value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
-                                                        case "resourcefunction":
-                                                            var functions = await ConfigurationHelper.GetList<ResourceFunction>();
-                                                            if (GeneralHelper.IsNotNull(functions))
-                                                            {
-                                                                var function = functions.Find(x => x.ShortName == value);
-                                                                if (!GeneralHelper.IsNotNull(function))
-                                                                {
-                                                                    valid = false;
-                                                                    sbMessage.Append("ResourceFunction value is invalid. ");
-                                                                }
-                                                            }
-                                                            break;
+                                                        sbMessage.Append(component.DisplayName + " value length is invalid. The value must be between " + component.MinLength + " and " + component.MaxLength + " characters.");
+                                                        valid = false;
                                                     }
-                                                    //var items = await ConfigurationHelper.GetList<ResourceComponent>();
 
-                                                    // Check if the delimeter is already ignored
-                                                    if ((!ignoredelimeter) && (!String.IsNullOrEmpty(resourceDelimiter.Delimiter)))
+                                                    if (valid)
                                                     {
-                                                        // Check if delimeter is an invalid character
-                                                        if (!String.IsNullOrEmpty(resourceType.InvalidCharacters))
+                                                        // Validate that the value is a valid option for the component
+                                                        switch (component.Name.ToLower())
                                                         {
-                                                            if (!resourceType.InvalidCharacters.Contains(resourceDelimiter.Delimiter))
-                                                            {
-                                                                if (name != "")
+                                                            case "resourcetype":
+                                                                var types = await ConfigurationHelper.GetList<ResourceType>();
+                                                                if (GeneralHelper.IsNotNull(types))
                                                                 {
-                                                                    name += resourceDelimiter.Delimiter;
+                                                                    var type = types.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(type))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceType value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourceenvironment":
+                                                                var environments = await ConfigurationHelper.GetList<ResourceEnvironment>();
+                                                                if (GeneralHelper.IsNotNull(environments))
+                                                                {
+                                                                    var environment = environments.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(environment))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceEnvironment value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourcelocation":
+                                                                var locations = await ConfigurationHelper.GetList<ResourceLocation>();
+                                                                if (GeneralHelper.IsNotNull(locations))
+                                                                {
+                                                                    var location = locations.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(location))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceLocation value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourceorg":
+                                                                var orgs = await ConfigurationHelper.GetList<ResourceOrg>();
+                                                                if (GeneralHelper.IsNotNull(orgs))
+                                                                {
+                                                                    var org = orgs.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(org))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("Resource Type value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourceprojappsvc":
+                                                                var projappsvcs = await ConfigurationHelper.GetList<ResourceProjAppSvc>();
+                                                                if (GeneralHelper.IsNotNull(projappsvcs))
+                                                                {
+                                                                    var projappsvc = projappsvcs.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(projappsvc))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceProjAppSvc value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourceunitdept":
+                                                                var unitdepts = await ConfigurationHelper.GetList<ResourceUnitDept>();
+                                                                if (GeneralHelper.IsNotNull(unitdepts))
+                                                                {
+                                                                    var unitdept = unitdepts.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(unitdept))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceUnitDept value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                            case "resourcefunction":
+                                                                var functions = await ConfigurationHelper.GetList<ResourceFunction>();
+                                                                if (GeneralHelper.IsNotNull(functions))
+                                                                {
+                                                                    var function = functions.Find(x => x.ShortName == value);
+                                                                    if (!GeneralHelper.IsNotNull(function))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append("ResourceFunction value is invalid. ");
+                                                                    }
+                                                                }
+                                                                break;
+                                                        }
+
+                                                        // Check if the delimiter is already ignored
+                                                        if ((!ignoredelimeter) && (!String.IsNullOrEmpty(resourceDelimiter.Delimiter)))
+                                                        {
+                                                            // Check if delimiter is an invalid character
+                                                            if (!String.IsNullOrEmpty(resourceType.InvalidCharacters))
+                                                            {
+                                                                if (!resourceType.InvalidCharacters.Contains(resourceDelimiter.Delimiter))
+                                                                {
+                                                                    if (name != "")
+                                                                    {
+                                                                        // Check if the component should apply the delimiter
+                                                                        if ((component.ApplyDelimiterBefore) && (previousdelimiterappliedafter) && (resourceType.ApplyDelimiter))
+                                                                        {
+                                                                            name += resourceDelimiter.Delimiter;
+                                                                        }
+                                                                    }
+                                                                }
+                                                                else
+                                                                {
+                                                                    // Add message about delimeter not applied
+                                                                    sbMessage.Append("The specified delimiter is not allowed for this resource type and has been removed. ");
+                                                                    ignoredelimeter = true;
                                                                 }
                                                             }
                                                             else
                                                             {
-                                                                // Add message about delimeter not applied
-                                                                sbMessage.Append("The specified delimiter is not allowed for this resource type and has been removed. ");
-                                                                ignoredelimeter = true;
+                                                                // Delimiter is valid so add it
+                                                                if (!String.IsNullOrEmpty(name))
+                                                                {
+                                                                    // Check if the component should apply the delimiter
+                                                                    if ((component.ApplyDelimiterBefore) && (previousdelimiterappliedafter) && (resourceType.ApplyDelimiter))
+                                                                    {
+                                                                        name += resourceDelimiter.Delimiter;
+                                                                    }
+                                                                }
                                                             }
                                                         }
-                                                        else
+
+                                                        name += value;
+
+                                                        // Add property to array for individual component validation
+                                                        if (!resourceType.Exclude.ToLower().Split(',').Contains(normalizedcomponentname))
                                                         {
-                                                            // Deliemeter is valid so add it
-                                                            if (!String.IsNullOrEmpty(name))
-                                                            {
-                                                                name += resourceDelimiter.Delimiter;
-                                                            }
+                                                            lstComponents.Add([component.Name, value]);
                                                         }
-                                                    }
-
-                                                    name += value;
-
-                                                    // Add property to array for individual component validation
-                                                    if (!resourceType.Exclude.ToLower().Split(',').Contains(normalizedcomponentname))
-                                                    {
-                                                        lstComponents.Add(new string[] { component.Name, value });
                                                     }
                                                 }
                                                 else
@@ -518,35 +549,36 @@ namespace AzureNamingTool.Services
                                             {
                                                 if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
                                                 {
-                                                    if (GeneralHelper.IsNotNull(serviceResponse.ResponseObject))
+                                                    var customcomponents = (List<CustomComponent>)serviceResponse.ResponseObject!;
+                                                    if (GeneralHelper.IsNotNull(customcomponents))
                                                     {
-                                                        var customcomponents = (List<CustomComponent>)serviceResponse.ResponseObject!;
-                                                        if (GeneralHelper.IsNotNull(customcomponents))
+                                                        // Make sure the custom component has values
+                                                        if (customcomponents.Where(x => x.ParentComponent == normalizedcomponentname).Any())
                                                         {
-                                                            // Make sure the custom component has values
-                                                            if (customcomponents.Where(x => x.ParentComponent == normalizedcomponentname).Any())
+                                                            // Make sure the CustomComponents property was provided
+                                                            if (!resourceType.Exclude.ToLower().Split(',').Contains(normalizedcomponentname))
                                                             {
-                                                                // Make sure the CustomComponents property was provided
-                                                                if (!resourceType.Exclude.ToLower().Split(',').Contains(normalizedcomponentname))
+                                                                // Add property value to name, if exists
+                                                                if (GeneralHelper.IsNotNull(request.CustomComponents))
                                                                 {
-                                                                    // Add property value to name, if exists
-                                                                    if (GeneralHelper.IsNotNull(request.CustomComponents))
+                                                                    // Check if the custom compoment value was provided in the request
+                                                                    if (request.CustomComponents.TryGetValue(normalizedcomponentname, out string? value))
                                                                     {
-                                                                        // Check if the custom compoment value was provided in the request
-                                                                        if (request.CustomComponents.ContainsKey(normalizedcomponentname))
+                                                                        // Get the value from the provided custom components
+                                                                        var componentvalue = value;
+                                                                        if (!GeneralHelper.IsNotNull(componentvalue))
                                                                         {
-                                                                            // Get the value from the provided custom components
-                                                                            var componentvalue = request.CustomComponents[normalizedcomponentname];
-                                                                            if (!GeneralHelper.IsNotNull(componentvalue))
+                                                                            // Check if the prop is optional
+                                                                            if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
                                                                             {
-                                                                                // Check if the prop is optional
-                                                                                if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
-                                                                                {
-                                                                                    valid = false;
-                                                                                    sbMessage.Append(component.Name + " value was not provided. ");
-                                                                                }
+                                                                                valid = false;
+                                                                                sbMessage.Append(component.Name + " value was not provided. ");
                                                                             }
-                                                                            else
+                                                                        }
+                                                                        else
+                                                                        {
+                                                                            // Check if the custom component is empty
+                                                                            if (!String.IsNullOrEmpty(componentvalue))
                                                                             {
                                                                                 // Check to make sure it is a valid custom component
                                                                                 var customComponents = await ConfigurationHelper.GetList<CustomComponent>();
@@ -560,7 +592,7 @@ namespace AzureNamingTool.Services
                                                                                     }
                                                                                     else
                                                                                     {
-                                                                                        if (!String.IsNullOrEmpty(name))
+                                                                                        if (!String.IsNullOrEmpty(name) && (resourceType.ApplyDelimiter))
                                                                                         {
                                                                                             name += resourceDelimiter.Delimiter;
                                                                                         }
@@ -568,18 +600,18 @@ namespace AzureNamingTool.Services
                                                                                         name += componentvalue;
 
                                                                                         // Add property to array for individual component validation
-                                                                                        lstComponents.Add(new string[] { component.Name, componentvalue });
+                                                                                        lstComponents.Add([component.Name, componentvalue]);
                                                                                     }
                                                                                 }
                                                                             }
-                                                                        }
-                                                                        else
-                                                                        {
-                                                                            // Check if the prop is optional
-                                                                            if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
+                                                                            else
                                                                             {
-                                                                                valid = false;
-                                                                                sbMessage.Append(component.Name + " value was not provided. ");
+                                                                                // Check if the prop is optional
+                                                                                if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
+                                                                                {
+                                                                                    valid = false;
+                                                                                    sbMessage.Append(component.Name + " value was not provided. ");
+                                                                                }
                                                                             }
                                                                         }
                                                                     }
@@ -593,9 +625,19 @@ namespace AzureNamingTool.Services
                                                                         }
                                                                     }
                                                                 }
+                                                                else
+                                                                {
+                                                                    // Check if the prop is optional
+                                                                    if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
+                                                                    {
+                                                                        valid = false;
+                                                                        sbMessage.Append(component.Name + " value was not provided. ");
+                                                                    }
+                                                                }
                                                             }
                                                         }
                                                     }
+
                                                 }
                                             }
                                         }
@@ -608,10 +650,10 @@ namespace AzureNamingTool.Services
                                                 if (GeneralHelper.IsNotNull(request.CustomComponents))
                                                 {
                                                     // Check if the custom compoment value was provided in the request
-                                                    if (request.CustomComponents.ContainsKey(normalizedcomponentname))
+                                                    if (request.CustomComponents.TryGetValue(normalizedcomponentname, out string? value))
                                                     {
                                                         // Get the value from the provided custom components
-                                                        var componentvalue = request.CustomComponents[normalizedcomponentname];
+                                                        var componentvalue = value;
                                                         if (!GeneralHelper.IsNotNull(componentvalue))
                                                         {
                                                             // Check if the prop is optional
@@ -623,15 +665,32 @@ namespace AzureNamingTool.Services
                                                         }
                                                         else
                                                         {
-                                                            if (!String.IsNullOrEmpty(name))
+                                                            // CHeck if the free text has a value
+                                                            if (!String.IsNullOrEmpty(componentvalue))
                                                             {
-                                                                name += resourceDelimiter.Delimiter;
+                                                                if (!String.IsNullOrEmpty(name))
+                                                                {
+                                                                    // Check if the component should apply the delimiter
+                                                                    if ((component.ApplyDelimiterBefore) && (previousdelimiterappliedafter) && (resourceType.ApplyDelimiter))
+                                                                    {
+                                                                        name += resourceDelimiter.Delimiter;
+                                                                    }
+                                                                }
+
+                                                                name += componentvalue;
+
+                                                                // Add property to array for individual component validation
+                                                                lstComponents.Add([component.Name, componentvalue]);
                                                             }
-
-                                                            name += componentvalue;
-
-                                                            // Add property to array for individual component validation
-                                                            lstComponents.Add(new string[] { component.Name, componentvalue });
+                                                            else
+                                                            {
+                                                                // Check if the prop is optional
+                                                                if (!resourceType.Optional.ToLower().Split(',').Contains(normalizedcomponentname))
+                                                                {
+                                                                    valid = false;
+                                                                    sbMessage.Append(component.Name + " value was not provided. ");
+                                                                }
+                                                            }
                                                         }
                                                     }
                                                     else
@@ -656,6 +715,8 @@ namespace AzureNamingTool.Services
                                             }
                                         }
                                     }
+                                    // Set the previous apply deimiter flag
+                                    previousdelimiterappliedafter = component.ApplyDelimiterAfter;
                                 }
                             }
                         }
@@ -695,6 +756,7 @@ namespace AzureNamingTool.Services
                 // Validate the generated name for the resource type
                 ValidateNameRequest validateNameRequest = new()
                 {
+                    ResourceTypeId = resourceType.Id,
                     ResourceType = resourceType.ShortName,
                     Name = name
                 };
@@ -725,31 +787,33 @@ namespace AzureNamingTool.Services
                         // Check if the request contains Resource Instance is a selected componoent
                         if (!String.IsNullOrEmpty(GeneralHelper.GetPropertyValue(request, "ResourceInstance")?.ToString()))
                         {
-                            // CHeck if the name should be auto-incremented
+                            // Check if the name should be auto-incremented
                             if (Convert.ToBoolean(ConfigurationHelper.GetAppSetting("AutoIncrementResourceInstance")))
                             {
                                 // Check if there was a ResourceInstance value supplied
                                 if (GeneralHelper.IsNotNull(GeneralHelper.GetPropertyValue(request, "ResourceInstance")))
                                 {
                                     // Attempt to auto-increement the instance 
+                                    // Set the original name value
+                                    string originalname = name;
                                     // Get the instance value
                                     string originalinstance = GeneralHelper.GetPropertyValue(request, "ResourceInstance")?.ToString() ?? "";
-                                    // Increase the instance by 1
-                                    string newinstance = (Convert.ToInt32(originalinstance) + 1).ToString();
-                                    // Make sure the instance pattern matches the entered values (leading zeros)
-                                    while (newinstance.Length < originalinstance.Length)
+                                    // Determine the next instance value
+                                    string newinstance = String.Empty;
+                                    int i = 1;
+                                    while (await ConfigurationHelper.CheckIfGeneratedNameExists(name))
                                     {
-                                        newinstance = "0" + newinstance;
+                                        newinstance = (Convert.ToInt32(originalinstance) + i).ToString();
+                                        // Make sure the instance pattern matches the entered values (leading zeros)
+                                        while (newinstance.Length < originalinstance.Length)
+                                        {
+                                            newinstance = "0" + newinstance;
+                                        }
+                                        // Replace the new instance in the original name
+                                        name = originalname.Replace(originalinstance, newinstance);
+                                        // Increase the counter
+                                        i += 1;
                                     }
-                                    // Update the generated name with the new instance
-                                    string newname = name.Replace(originalinstance, newinstance);
-                                    // Check to make sure the new name is unique
-                                    while (await ConfigurationHelper.CheckIfGeneratedNameExists(newname))
-                                    {
-                                        newinstance = (Convert.ToInt32(newinstance) + 1).ToString();
-                                        newname = name.Replace(originalinstance, newinstance);
-                                    }
-                                    name = newname;
                                     sbMessage.Append("The resource instance has been auto-incremented to the next value.");
                                     sbMessage.Append(Environment.NewLine);
                                 }
@@ -787,7 +851,7 @@ namespace AzureNamingTool.Services
                         GeneratedName generatedName = new()
                         {
                             CreatedOn = DateTime.Now,
-                            ResourceName = name.ToLower(),
+                            ResourceName = name,
                             Components = lstComponents,
                             ResourceTypeName = resourceType.Resource,
                             User = request.CreatedBy,
@@ -804,7 +868,7 @@ namespace AzureNamingTool.Services
                         if (responseGenerateName.Success)
                         {
                             resourceNameResponse.Success = true;
-                            resourceNameResponse.ResourceName = name.ToLower();
+                            resourceNameResponse.ResourceName = name;
                             resourceNameResponse.Message = sbMessage.ToString();
                             resourceNameResponse.ResourceNameDetails = generatedName;
 
